Refactor most request workflow step modules
