+++ trustpoint/request/authentication/__init__.py
@@ -1,1 +1,19 @@
 """Initialization for the authentication step of the request pipeline."""
+
+from .base import (
+    AuthenticationComponent,
+    ClientCertificateAuthentication,
+    CompositeAuthentication,
+    IDevIDAuthentication,
+)
+from .cmp import CmpAuthentication
+from .est import EstAuthentication
+
+__all__ = [
+    'AuthenticationComponent',
+    'ClientCertificateAuthentication',
+    'CmpAuthentication',
+    'CompositeAuthentication',
+    'EstAuthentication',
+    'IDevIDAuthentication',
+]
