+++ trustpoint/request/authentication/cmp.py
@@ -1,6 +1,5 @@
-"""Provides the `EstAuthentication` class using the Composite pattern for modular EST authentication."""
+"""Provides the 'CmpAuthentication' class using the Composite pattern for modular CMP authentication."""
 
-from abc import ABC, abstractmethod
 from typing import Never
 
 from cryptography import x509
@@ -13,280 +12,22 @@ from trustpoint_core.oid import AlgorithmIdentifier, HashAlgorithm, HmacAlgorith
 
 from devices.models import (
     DeviceModel,
-    IssuedCredentialModel,
     NoOnboardingConfigModel,
     OnboardingConfigModel,
     OnboardingPkiProtocol,
     OnboardingProtocol,
 )
-from pki.models import CredentialModel
-from pki.util.idevid import IDevIDAuthenticationError, IDevIDAuthenticator
-from request.request_context import RequestContext
+from pki.util.idevid import IDevIDAuthenticator
+from request.request_context import CmpBaseRequestContext, CmpCertificateRequestContext
 from trustpoint.logger import LoggerMixin
 
+from .base import AuthenticationComponent, ClientCertificateAuthentication, CompositeAuthentication
 
-class AuthenticationComponent(ABC):
-    """Abstract base class for authentication components."""
-
-    @abstractmethod
-    def authenticate(self, context: RequestContext) -> DeviceModel | None:
-        """Authenticate a request using specific logic."""
-
-
-class UsernamePasswordAuthentication(AuthenticationComponent, LoggerMixin):
-    """Handles authentication via username/password credentials."""
-
-    def authenticate(self, context: RequestContext) -> None:
-        """Authenticate using username and password from the context."""
-        if not (context.est_username and context.est_password):
-            return
-
-        username = context.est_username
-        password = context.est_password
-
-        try:
-            device = DeviceModel.objects.select_related().filter(
-                common_name=username
-            ).first()
-
-            if not device:
-                self.logger.warning('Authentication failed: Unknown username %s', username)
-                self._raise_authentication_error()
-
-            device_config = device.onboarding_config or device.no_onboarding_config
-
-            if not isinstance(device, DeviceModel) or not device_config:
-                self.logger.warning('Authentication failed: Invalid device model for %s', username)
-                self._raise_authentication_error()
-
-            if not device_config.est_password:
-                self.logger.warning('Authentication failed: No EST password set for %s', username)
-                self._raise_authentication_error()
-
-            # Use proper password hashing instead of plaintext comparison
-            if password != device_config.est_password:
-                self.logger.warning('Authentication failed: Invalid password for %s', username)
-                self._raise_authentication_error()
-
-            self.logger.info('Successfully authenticated device %s', username)
-            context.device = device
-
-        except Exception as e:
-            self.logger.exception('Authentication error for user %s', username)
-            error_message = 'Authentication failed: Invalid username or password.'
-            raise ValueError(error_message) from e
-
-    def _raise_authentication_error(self) -> Never:
-        """Raise authentication error with standardized message."""
-        error_message = 'Authentication failed: Invalid username or password.'
-        raise ValueError(error_message)
-
-
-class ClientCertificateAuthentication(AuthenticationComponent, LoggerMixin):
-    """Handles authentication via client certificates."""
-
-    def authenticate(self, context: RequestContext) -> None:
-        """Authenticate using the client certificate from the context."""
-        if not context.client_certificate:
-            return
-
-        client_certificate = context.client_certificate
-
-        try:
-            issued_credential = IssuedCredentialModel.get_credential_for_certificate(client_certificate)
-
-            is_valid, reason = issued_credential.is_valid_domain_credential()
-            if not is_valid:
-                self.logger.warning('Invalid client certificate: %s', reason)
-                error_message = f'Invalid HTTP_SSL_CLIENT_CERT header: {reason}'
-                self._raise_certificate_error(error_message)
-
-            self.logger.info('Successfully authenticated device via client certificate')
-            context.device = issued_credential.device
-
-        except IssuedCredentialModel.DoesNotExist:
-            self.logger.warning('Client certificate not found in issued credentials')
-            error_message = 'Client certificate not recognized'
-            self._raise_certificate_error(error_message)
-        except ValueError:
-            raise
-        except Exception as e:
-            self.logger.exception('Certificate authentication error')
-            error_message = 'Certificate authentication failed'
-            self._raise_certificate_error(error_message, e)
-
-    def _raise_certificate_error(self, message: str, cause: Exception | None = None) -> Never:
-        """Raise certificate authentication error with proper exception chaining."""
-        if cause:
-            raise ValueError(message) from cause
-        raise ValueError(message)
-
-
-class ReenrollmentAuthentication(AuthenticationComponent, LoggerMixin):
-    """Handles authentication for EST reenrollment using an Application Credential."""
-
-    def _validate_certificate_extensions(
-        self,
-        credential_cert: x509.Certificate,
-        client_cert: x509.Certificate,
-        csr: x509.CertificateSigningRequest
-    ) -> None:
-        """Validate that certificate extensions match between credential, client cert, and CSR."""
-        try:
-            credential_cert_san = credential_cert.extensions.get_extension_for_class(x509.SubjectAlternativeName)
-        except x509.ExtensionNotFound:
-            credential_cert_san = None
-
-        try:
-            csr_san = csr.extensions.get_extension_for_class(x509.SubjectAlternativeName)
-        except x509.ExtensionNotFound:
-            csr_san = None
-
-        try:
-            client_san = client_cert.extensions.get_extension_for_class(x509.SubjectAlternativeName)
-        except x509.ExtensionNotFound:
-            client_san = None
-
-        if client_san != csr_san or credential_cert_san != csr_san:
-            error_message = 'CSR/client SAN does not match the credential certificate SAN.'
-            raise ValueError(error_message)
-
-    def _raise_value_error(self, message: str) -> Never:
-        """Raise a ValueError with the given message."""
-        raise ValueError(message)
-
-    def authenticate(self, context: RequestContext) -> None:
-        """Authenticate the client for reenrollment."""
-        if not self._validate_context(context):
-            return
-
-        if not context.client_certificate:
-            error_message = 'Client certificate is required for reenrollment.'
-            self.logger.warning(error_message)
-            self._raise_value_error(error_message)
-
-        issued_credential = self._get_issued_credential(context.client_certificate)
-        credential_model: CredentialModel = issued_credential.credential
-
-        if not isinstance(context.cert_requested, x509.CertificateSigningRequest):
-            error_message = 'Invalid credential model for reenrollment.'
-            self.logger.warning(error_message)
-            self._raise_value_error(error_message)
-
-        self._validate_credential(credential_model, context.cert_requested, context.client_certificate)
-        self._validate_certificate_extensions_safe(credential_model, context.client_certificate, context.cert_requested)
-
-        self.logger.info('Successfully authenticated device for reenrollment')
-        context.device = issued_credential.device
-
-    def _validate_context(self, context: RequestContext) -> bool:
-        """Validate the context for reenrollment."""
-        if not context.client_certificate:
-            return False
-
-        if not isinstance(context.client_certificate, x509.Certificate):
-            error_message = 'Invalid client certificate type for reenrollment.'
-            self.logger.warning(error_message)
-            self._raise_value_error(error_message)
-
-        if not context.cert_requested:
-            error_message = 'CSR is missing in the context for reenrollment.'
-            self.logger.warning(error_message)
-            self._raise_value_error(error_message)
-
-        return True
-
-    def _get_issued_credential(self, client_cert: x509.Certificate) -> IssuedCredentialModel:
-        """Retrieve the issued credential for the client certificate."""
-        try:
-            return IssuedCredentialModel.get_credential_for_certificate(client_cert)
-        except IssuedCredentialModel.DoesNotExist:
-            error_message = 'Issued credential not found for client certificate during reenrollment'
-            self.logger.warning(error_message)
-            raise ValueError(error_message) from None
-
-    def _validate_credential(
-        self, credential_model: CredentialModel, csr: x509.CertificateSigningRequest, client_cert: x509.Certificate
-    ) -> None:
-        """Validate the credential model against the CSR and client certificate."""
-        is_valid, reason = credential_model.is_valid_issued_credential()
-        if not is_valid:
-            error_message = f'Invalid client certificate for reenrollment: {reason}'
-            self.logger.warning(error_message)
-            raise ValueError(error_message)
-
-        if (
-            not credential_model.certificate.subjects_match(csr.subject) or
-            not credential_model.certificate.subjects_match(client_cert.subject)
-        ):
-            error_message = "CSR/client subject does not match the credential certificate's subject"
-            self.logger.warning(error_message)
-            raise ValueError(error_message)
-
-    def _validate_certificate_extensions_safe(
-        self, credential_model: CredentialModel, client_cert: x509.Certificate, csr: x509.CertificateSigningRequest
-    ) -> None:
-        """Safely validate certificate extensions."""
-        try:
-            credential_cert = credential_model.certificate.get_certificate_serializer().as_crypto()
-            self._validate_certificate_extensions(credential_cert, client_cert, csr)
-        except Exception as e:
-            self.logger.warning('Certificate extension validation failed: %s', e)
-            error_message = 'Certificate extension validation failed'
-            raise ValueError(error_message) from e
-
-
-class IDevIDAuthentication(AuthenticationComponent, LoggerMixin):
-    """Handles authentication via IDevID certificates."""
-
-    def authenticate(self, context: RequestContext) -> None:
-        """Authenticate the request using the IDevID mechanism."""
-        # Early return if raw_message is missing
-        if not context.raw_message:
-            return
-
-        try:
-            # domain can be None, in that case IDevIDAuthenticator will infer it from registration pattern
-            device_or_none = IDevIDAuthenticator.authenticate_idevid(context.raw_message, context.domain)
-
-            if device_or_none:
-                self.logger.info('Successfully authenticated device via IDevID')
-                context.device = device_or_none
-                if not context.domain:
-                    device_domain = context.device.domain
-                    if not device_domain:
-                        error_message = 'IDevID authentication failed: Device domain is not set.'
-                        self.logger.warning('IDevID authentication failed: Device domain is not set')
-                        self._raise_idevid_error(error_message)
-                    context.domain = device_domain
-                    context.domain_str = device_domain.unique_name
-            else:
-                error_message = 'IDevID authentication failed: No device associated.'
-                self.logger.warning('IDevID authentication failed: No device associated')
-                self._raise_idevid_error(error_message)
-
-        except IDevIDAuthenticationError as e:
-            error_message = f'Error validating the IDevID: {e}'
-            self.logger.warning('Error validating the IDevID: %s', e)
-            raise ValueError(error_message) from e
-        except ValueError:
-            raise
-        except Exception as e:
-            error_message = 'IDevID authentication failed due to unexpected error'
-            self.logger.exception('Unexpected error during IDevID authentication')
-            raise ValueError(error_message) from e
-
-    def _raise_idevid_error(self, message: str) -> Never:
-        """Raise IDevID authentication error."""
-        raise ValueError(message)
-
-
-#####
 
 class CmpAuthenticationBase(AuthenticationComponent, LoggerMixin):
     """Base class for CMP authentication components with common functionality."""
 
-    def _is_aoki_request(self, context: RequestContext) -> bool:
+    def _is_aoki_request(self, context: CmpBaseRequestContext) -> bool:
         """Determine if this is an AOKI request based on domain name and URL path."""
         domain_name = context.domain_str
         request_path = getattr(context, 'request_path', None)
@@ -300,7 +41,7 @@ class CmpAuthenticationBase(AuthenticationComponent, LoggerMixin):
 class CmpSharedSecretAuthentication(CmpAuthenticationBase):
     """Handles CMP authentication using shared secrets with HMAC-based protection."""
 
-    def authenticate(self, context: RequestContext) -> None:
+    def authenticate(self, context: CmpBaseRequestContext) -> None:
         """Authenticate using CMP shared secret HMAC protection."""
         if not self._validate_context(context):
             return
@@ -317,7 +58,7 @@ class CmpSharedSecretAuthentication(CmpAuthenticationBase):
         except Exception as e:  # noqa: BLE001
             self._handle_unexpected_error(e)
 
-    def _validate_context(self, context: RequestContext) -> bool:
+    def _validate_context(self, context: CmpBaseRequestContext) -> bool:
         """Validate the context for CMP shared secret authentication."""
         if context.protocol != 'cmp':
             error_message = 'CMP shared secret authentication requires CMP protocol.'
@@ -354,7 +95,7 @@ class CmpSharedSecretAuthentication(CmpAuthenticationBase):
         self.logger.warning(message)
         raise ValueError(message)
 
-    def _extract_sender_kid(self, context: RequestContext) -> int:
+    def _extract_sender_kid(self, context: CmpBaseRequestContext) -> int:
         """Extract sender KID from CMP message header."""
         try:
             if not isinstance(context.parsed_message, rfc4210.PKIMessage):
@@ -387,12 +128,12 @@ class CmpSharedSecretAuthentication(CmpAuthenticationBase):
             self._raise_cmp_error(error_message)
         return device_config
 
-    def _verify_hmac_protection(self, context: RequestContext, shared_secret: str) -> None:
+    def _verify_hmac_protection(self, context: CmpBaseRequestContext, shared_secret: str) -> None:
         """Verify HMAC-based protection and store shared secret for response."""
         self._verify_protection_shared_secret(context.parsed_message, shared_secret)
         context.cmp_shared_secret = shared_secret
 
-    def _finalize_authentication(self, context: RequestContext, device: DeviceModel, sender_kid: int) -> None:
+    def _finalize_authentication(self, context: CmpBaseRequestContext, device: DeviceModel, sender_kid: int) -> None:
         """Finalize authentication by setting device in context and logging success."""
         self.logger.info(
             'Successfully authenticated device %s (ID: %s) via CMP shared secret', device.common_name, sender_kid)
@@ -479,7 +220,7 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
         """Initialize the CMP signature-based authentication component."""
 
 
-    def authenticate(self, context: RequestContext) -> None:
+    def authenticate(self, context: CmpCertificateRequestContext) -> None:
         """Authenticate using CMP signature-based protection for initialization requests."""
         if not self._validate_context(context):
             return
@@ -492,7 +233,7 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
         context.device = device
 
     def _authenticate_and_verify_device(self,
-                                        context: RequestContext,
+                                        context: CmpCertificateRequestContext,
                                         cmp_signer_cert: x509.Certificate,
                                         intermediate_certs: list[x509.Certificate]) -> DeviceModel:
         """Authenticate and verify the device."""
@@ -504,7 +245,8 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
             return device
 
     def _process_device_authentication(
-        self, context: RequestContext, cmp_signer_cert: x509.Certificate, intermediate_certs: list[x509.Certificate]
+        self, context: CmpCertificateRequestContext,
+        cmp_signer_cert: x509.Certificate, intermediate_certs: list[x509.Certificate]
     ) -> DeviceModel:
         """Process device authentication using certificates."""
         device = self._authenticate_device(context, cmp_signer_cert, intermediate_certs)
@@ -518,7 +260,7 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
         self.logger.warning(error_message)
         raise ValueError(error_message) from error
 
-    def _validate_context(self, context: RequestContext) -> bool:
+    def _validate_context(self, context: CmpCertificateRequestContext) -> bool:
         """Validate the context for CMP authentication."""
         if context.protocol != 'cmp':
             self._raise_value_error('CMP shared secret authentication requires CMP protocol.')
@@ -534,7 +276,8 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
 
         return True
 
-    def _extract_certificates(self, context: RequestContext) -> tuple[x509.Certificate, list[x509.Certificate]]:
+    def _extract_certificates(
+            self, context: CmpCertificateRequestContext) -> tuple[x509.Certificate, list[x509.Certificate]]:
         """Extract and validate certificates from the CMP message."""
         if not context or not context.parsed_message:
             err_msg = 'Missing parsed message in context.'
@@ -563,7 +306,7 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
 
         return cmp_signer_cert, intermediate_certs
 
-    def _authenticate_device(self, context: RequestContext, cmp_signer_cert: x509.Certificate,
+    def _authenticate_device(self, context: CmpCertificateRequestContext, cmp_signer_cert: x509.Certificate,
                              intermediate_certs: list[x509.Certificate]) -> DeviceModel:
         """Authenticate the device using IDevID."""
         is_aoki = self._is_aoki_request(context)
@@ -637,7 +380,7 @@ class CmpSignatureBasedInitializationAuthentication(CmpAuthenticationBase):
 class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, LoggerMixin):
     """Handles CMP signature-based authentication for certification requests using domain credentials."""
 
-    def authenticate(self, context: RequestContext) -> None:
+    def authenticate(self, context: CmpCertificateRequestContext) -> None:
         """Authenticate using CMP signature-based protection for certification requests."""
         if not self._should_authenticate(context):
             return
@@ -652,7 +395,7 @@ class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, Logg
             self.logger.warning(error_message)
             raise ValueError(error_message) from e
 
-    def _should_authenticate(self, context: RequestContext) -> bool:
+    def _should_authenticate(self, context: CmpCertificateRequestContext) -> bool:
         """Check if this authentication method should be applied."""
         if context.protocol != 'cmp':
             return False
@@ -688,7 +431,7 @@ class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, Logg
 
         return True
 
-    def _extract_and_validate_certificate(self, context: RequestContext) -> x509.Certificate:
+    def _extract_and_validate_certificate(self, context: CmpCertificateRequestContext) -> x509.Certificate:
         """Extract and validate the CMP signer certificate from the message."""
         if not context or not context.parsed_message:
             err_msg = 'Missing parsed message in context.'
@@ -710,7 +453,7 @@ class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, Logg
         context.client_certificate = x509.load_der_x509_certificate(der_cmp_signer_cert)
         return context.client_certificate
 
-    def _authenticate_device(self, context: RequestContext) -> DeviceModel:
+    def _authenticate_device(self, context: CmpCertificateRequestContext) -> DeviceModel:
         """Authenticate the device using the CMP signer certificate."""
         cmp_signer_cert = context.client_certificate
         if not cmp_signer_cert:
@@ -835,7 +578,7 @@ class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, Logg
             self._raise_value_error(error_message)
 
     def _verify_protection_and_finalize(
-        self, context: RequestContext, cmp_signer_cert: x509.Certificate, device: DeviceModel
+        self, context: CmpCertificateRequestContext, cmp_signer_cert: x509.Certificate, device: DeviceModel
     ) -> None:
         """Verify protection signature and finalize authentication."""
         # Verify protection signature
@@ -897,56 +640,6 @@ class CmpSignatureBasedCertificationAuthentication(AuthenticationComponent, Logg
             err_msg = 'Cannot verify signature due to unsupported public key type.'
             raise TypeError(err_msg)
 
-class CompositeAuthentication(AuthenticationComponent, LoggerMixin):
-    """Composite authenticator for grouping and executing multiple authentication methods."""
-
-    def __init__(self) -> None:
-        """Initialize the composite authenticator with a set of authentication components."""
-        self.components: list[AuthenticationComponent] = []
-
-    def add(self, component: AuthenticationComponent) -> None:
-        """Add an authentication component to the composite."""
-        self.components.append(component)
-
-    def remove(self, component: AuthenticationComponent) -> None:
-        """Remove an authentication component from the composite."""
-        self.components.remove(component)
-
-    def authenticate(self, context: RequestContext) -> None:
-        """Authenticate the request using all registered components."""
-        authentication_errors = []
-
-        self.logger.debug('Authentication components to try: %s',
-                          [component.__class__.__name__ for component in self.components])
-        for component in self.components:
-            try:
-                component.authenticate(context)
-                if context.device is not None:
-                    self.logger.info('Authentication successful using %s', component.__class__.__name__)
-                    return
-            except ValueError as e:
-                authentication_errors.append(f'{component.__class__.__name__}: {e}')
-                continue
-            except Exception:
-                self.logger.exception('Unexpected error in %s', component.__class__.__name__)
-                authentication_errors.append(f'{component.__class__.__name__}: Unexpected error')
-                continue
-        error_message = 'Authentication failed: All authentication methods were unsuccessful.'
-        self.logger.warning('Authentication failed for all methods: %s', authentication_errors)
-        context.http_response_content = 'Authentication failed.'
-        context.http_response_status = 403
-        raise ValueError(error_message)
-
-class EstAuthentication(CompositeAuthentication):
-    """Composite authenticator specifically for EST requests, combining various authentication methods."""
-
-    def __init__(self) -> None:
-        """Initialize the EST authenticator with a set of authentication methods."""
-        super().__init__()
-        self.add(ReenrollmentAuthentication())
-        self.add(UsernamePasswordAuthentication())
-        self.add(ClientCertificateAuthentication())
-        self.add(IDevIDAuthentication())
 
 class CmpAuthentication(CompositeAuthentication):
     """Composite authenticator specifically for CMP requests, combining various authentication methods."""
@@ -957,4 +650,3 @@ class CmpAuthentication(CompositeAuthentication):
         self.add(CmpSharedSecretAuthentication())
         self.add(CmpSignatureBasedInitializationAuthentication())
         self.add(CmpSignatureBasedCertificationAuthentication())
-
