+++ trustpoint/request/authorization/__init__.py
@@ -1,1 +1,23 @@
 """Initialization for the authorization step of the request pipeline."""
+
+from .base import (
+    AuthorizationComponent,
+    CertificateProfileAuthorization,
+    CompositeAuthorization,
+    DevOwnerIDAuthorization,
+    DomainScopeValidation,
+    ProtocolAuthorization,
+)
+from .cmp import CmpAuthorization
+from .est import EstAuthorization
+
+__all__ = [
+    'AuthorizationComponent',
+    'CertificateProfileAuthorization',
+    'CmpAuthorization',
+    'CompositeAuthorization',
+    'DevOwnerIDAuthorization',
+    'DomainScopeValidation',
+    'EstAuthorization',
+    'ProtocolAuthorization',
+]
