+++ trustpoint/request/authorization/base.py
@@ -1,11 +1,8 @@
 """Provides the `AuthorizationComponent` class for authorization logic."""
 from abc import ABC, abstractmethod
-from typing import Never
-
-from pyasn1_modules.rfc4210 import PKIMessage  # type: ignore[import-untyped]
 
 from aoki.views import AokiServiceMixin
-from request.request_context import RequestContext
+from request.request_context import BaseCertificateRequestContext, BaseRequestContext
 from trustpoint.logger import LoggerMixin
 
 
@@ -13,7 +10,7 @@ class AuthorizationComponent(ABC):
     """Abstract base class for authorization components."""
 
     @abstractmethod
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseRequestContext) -> None:
         """Execute authorization logic."""
 
 class ProtocolAuthorization(AuthorizationComponent, LoggerMixin):
@@ -23,7 +20,7 @@ class ProtocolAuthorization(AuthorizationComponent, LoggerMixin):
         """Initialize the authorization component with a list of allowed protocols."""
         self.allowed_protocols = allowed_protocols
 
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseRequestContext) -> None:
         """Authorize the request based on the protocol."""
         protocol = context.protocol
 
@@ -45,106 +42,11 @@ class ProtocolAuthorization(AuthorizationComponent, LoggerMixin):
         self.logger.debug('Protocol authorization successful for protocol: %(protocol)s',
                           extra={'protocol': protocol})
 
-class EstOperationAuthorization(AuthorizationComponent, LoggerMixin):
-    """Ensures the request is authorized for the specified operation."""
-
-    def __init__(self, allowed_operations: list[str]) -> None:
-        """Initialize the authorization component with a list of allowed operations."""
-        self.allowed_operations = allowed_operations
-
-    def authorize(self, context: RequestContext) -> None:
-        """Authorize the request based on the operation type."""
-        operation = context.operation
-
-        if not operation:
-            error_message = 'Operation information is missing. Authorization denied.'
-            self.logger.warning('Operation authorization failed: Operation information is missing')
-            raise ValueError(error_message)
-
-        if operation not in self.allowed_operations:
-            error_message = (
-                f"Unauthorized operation: '{operation}'. "
-                f"Allowed operations: {', '.join(self.allowed_operations)}."
-            )
-            self.logger.warning(
-                'Operation authorization failed: %s not in allowed operations %s',
-                operation, self.allowed_operations
-            )
-            raise ValueError(error_message)
-
-        self.logger.debug('Operation authorization successful for operation: %s', operation)
-
-class CmpOperationAuthorization(AuthorizationComponent, LoggerMixin):
-    """Ensures the request is authorized for the specified operation."""
-
-    def __init__(self, allowed_operations: list[str]) -> None:
-        """Initialize the authorization component with a list of allowed operations."""
-        self.allowed_operations = allowed_operations
-
-    def authorize(self, context: RequestContext) -> None:
-        """Authorize the request based on the operation type."""
-        operation = context.operation
-
-        if not operation:
-            error_message = 'Operation information is missing. Authorization denied.'
-            self.logger.warning('Operation authorization failed: Operation information is missing')
-            raise ValueError(error_message)
-
-        if operation not in self.allowed_operations:
-            error_message = (
-                f"Unauthorized operation: '{operation}'. "
-                f"Allowed operations: {', '.join(self.allowed_operations)}."
-            )
-            self.logger.warning(
-                'Operation authorization failed: %(operation)s not in allowed operations %(allowed_operations)s',
-                extra={'operation': operation, 'allowed_operations': self.allowed_operations})
-            raise ValueError(error_message)
-
-        if not isinstance(context.parsed_message, PKIMessage):
-            error_message = 'Parsed message is missing. Authorization denied.'
-            self.logger.warning('Operation authorization failed: Parsed message is missing or invalid')
-            self._raise_value_error(error_message)
-
-        body_type = context.parsed_message['body'].getName()
-
-        if context.operation == 'initialization' and body_type == 'ir':
-            self._authorize_asn1_body(context.parsed_message, 'ir')
-            self.logger.info('CMP body type validation successful: IR body extracted')
-        elif context.operation == 'certification' and body_type == 'cr':
-            self._authorize_asn1_body(context.parsed_message, 'cr')
-            self.logger.info('CMP body type validation successful: CR body extracted')
-        else:
-            err_msg = f'Expected CMP {context.operation} body, but got CMP {body_type.upper()} body.'
-            raise ValueError(err_msg)
-
-        self.logger.debug('Operation authorization successful for operation: %(operation)s',
-                          extra={'operation': operation})
-
-    def _raise_value_error(self, message: str) -> Never:
-        """Raise a ValueError with the given message."""
-        raise ValueError(message)
-
-    def _authorize_asn1_body(self, serialized_pyasn1_message: PKIMessage, expected_body_type: str) -> None:
-        """Extract and validate the specified body type from the CMP message.
-
-        Args:
-            serialized_pyasn1_message: The CMP message to extract the body from.
-            expected_body_type: The expected body type ('cr' or 'ir').
-        """
-        message_body_name = serialized_pyasn1_message['body'].getName()
-        if message_body_name != expected_body_type:
-            err_msg = f'Expected CMP {expected_body_type.upper()} body, but got CMP {message_body_name.upper()} body.'
-            raise ValueError(err_msg)
-
-        if serialized_pyasn1_message['body'].getName() != expected_body_type:
-            err_msg = f'not {expected_body_type} message'
-            raise ValueError(err_msg)
-
 
 class CertificateProfileAuthorization(AuthorizationComponent, LoggerMixin):
     """Ensures the device is allowed to use the requested certificate profile."""
 
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseCertificateRequestContext) -> None:
         """Authorize the request based on the certificate profile."""
         requested_profile = context.cert_profile_str
 
@@ -179,7 +81,7 @@ class CertificateProfileAuthorization(AuthorizationComponent, LoggerMixin):
 class DomainScopeValidation(AuthorizationComponent, LoggerMixin):
     """Ensures the request is within the authorized domain."""
 
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseRequestContext) -> None:
         """Authorize the request based on the domain scope."""
         authenticated_device = context.device
         requested_domain = context.domain
@@ -216,7 +118,7 @@ class DomainScopeValidation(AuthorizationComponent, LoggerMixin):
 class DevOwnerIDAuthorization(AuthorizationComponent, LoggerMixin):
     """Ensure that if this is an AOKI request, we have a matching DevOwnerID to the IDevID."""
 
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseRequestContext) -> None:
         """Authorize the request based on the DevOwnerID corresponding to the client certificate."""
         if context.protocol != 'cmp':
             return
@@ -241,17 +143,6 @@ class DevOwnerIDAuthorization(AuthorizationComponent, LoggerMixin):
         context.owner_credential = owner_credential
 
 
-
-class ManualAuthorization(AuthorizationComponent, LoggerMixin):
-    """Perform manual authorization override."""
-
-    def authorize(self, context: RequestContext) -> None:
-        """Authorize the request based on manual authorization."""
-        del context
-        self.logger.debug('Manual authorization check passed (placeholder implementation)')
-
-
-
 class CompositeAuthorization(AuthorizationComponent, LoggerMixin):
     """Composite authorization handler for grouping and executing multiple authorization components."""
 
@@ -273,7 +164,7 @@ class CompositeAuthorization(AuthorizationComponent, LoggerMixin):
             self.logger.warning(error_message)
             raise ValueError(error_message)
 
-    def authorize(self, context: RequestContext) -> None:
+    def authorize(self, context: BaseRequestContext) -> None:
         """Iterate through all child authorization components and execute their authorization logic."""
         self.logger.debug('Starting composite authorization with %d components', len(self.components))
 
@@ -302,43 +193,3 @@ class CompositeAuthorization(AuthorizationComponent, LoggerMixin):
                 raise ValueError(error_message) from e
 
         self.logger.info('Composite authorization successful. All %d components passed', len(self.components))
-
-
-class EstAuthorization(CompositeAuthorization):
-    """Composite authorization handler for EST requests."""
-    def __init__(self, allowed_operations: list[str] | None = None) -> None:
-        """Initialize the composite authorization handler with the default set of components.
-
-        Args:
-            allowed_operations: List of allowed CMP operations. Defaults to ['cr', 'ir'] if not provided.
-        """
-        super().__init__()
-
-        if allowed_operations is None:
-            allowed_operations = ['simpleenroll', 'simplereenroll']
-
-        self.add(DomainScopeValidation())
-        self.add(CertificateProfileAuthorization())
-        self.add(ManualAuthorization())
-        self.add(ProtocolAuthorization(['est']))
-        self.add(EstOperationAuthorization(allowed_operations))
-
-class CmpAuthorization(CompositeAuthorization):
-    """Composite authorization handler for EST requests."""
-    def __init__(self, allowed_operations: list[str] | None = None) -> None:
-        """Initialize the composite authorization handler with the default set of components.
-
-        Args:
-            allowed_operations: List of allowed CMP operations. Defaults to ['cr', 'ir'] if not provided.
-        """
-        super().__init__()
-
-        if allowed_operations is None:
-            allowed_operations = ['certification', 'initialization']
-
-        self.add(CertificateProfileAuthorization())
-        self.add(DomainScopeValidation())
-        self.add(DevOwnerIDAuthorization())
-        self.add(ManualAuthorization())
-        self.add(ProtocolAuthorization(['cmp']))
-        self.add(CmpOperationAuthorization(allowed_operations))
