+++ trustpoint/request/message_parser/__init__.py
@@ -1,1 +1,9 @@
 """Initialization for the message parsing step of the request pipeline."""
+
+from .cmp import CmpMessageParser
+from .est import EstMessageParser
+
+__all__ = [
+    'CmpMessageParser',
+    'EstMessageParser',
+]
