+++ trustpoint/request/operation_processor/__init__.py
@@ -1,1 +1,9 @@
 """Initialization for the operation processing step of the request pipeline."""
+
+from .issue_cert import CertificateIssueProcessor
+from .sign import LocalCaCmpSignatureProcessor
+
+__all__ = [
+    'CertificateIssueProcessor',
+    'LocalCaCmpSignatureProcessor',
+]
