+++ trustpoint/request/tests/test_http_request_validator.py
@@ -3,7 +3,7 @@
 import base64
 from unittest.mock import MagicMock
 
-from request.http_request_validator import (
+from request.request_validator.http_request_validator import (
     AcceptHeaderValidation,
     AuthorizationHeaderValidation,
     ClientCertificateValidation,
